+++ marko/__init__.py
@@ -19,7 +19,7 @@ if is_type_check():
     from .block import Document
     from .parser import ElementType
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 
 class SetupDone(Exception):